Add better way to prevent div-by-zero in hcox_gc_RnPbBe_mod.F90
The variables F_BELOW_70, F_BELOW_60, F_ABOVE_60 were computed for
every grid box, but they are only used for a few grid boxes depending
on latitude.  We now define these variables in the IF block where
they are used.  This prevents the computation of bad values or
div-by-zero values, as well as being more computationaly efficient.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>