- Allow extension number of base emissions to be set to wildcard character. HEMCO will read those fields (if the assigned species is a valid HEMCO species) but not use them for emission calculation. This is particularly useful for data that is used outside of HEMCO, e.g. stratospheric prod/loss rates.
- Update of the range attribute (CRE) set in the HEMCO configuration file. If
a range is given, use the data container if the simulation date is
within the provided range, even if the input data does not cover the current
date. For instance, if a netCDF file covers years 2003 to 2010 and the range
is set to 2008-2100/1/1/0, the container will only be considered for simulation
years 2008 and higher. For year 2008 through 2010, the corresponding time slices
will be used. For years 2011 and higher, data of 2010 will be used.