Now avoid array temporary message in HEMCO output
When we run HEMCO with the DEBUG=y flag turned on, we get this output:

  forrtl: warning (402): fort: (1): In call to
  M_NETCDF_IO_WRITE^NCWR_1D_R8, an array temporary was created
  for argument #1

This is being caused at line 262 of HEMCO/Core/hcoio_diagn_mod.F90

    ! Add latitude
    CALL NC_VAR_DEF ( fId, -1, latId, -1, -1, &
                      'lat', 'Latitude', 'degrees_north', Prc, VarCt )
    Arr1D => HcoState%Grid%YMID%Val(1,:)
    CALL NC_VAR_WRITE ( fId, 'lat', Arr1D=Arr1D )
    Arr1D => NULL()

When we point to the latitude (in HcoState%GRID%Arr%Val(:,1)) we are
pointing to data that are non-contiguous in memory.  So the code has to
do an internal copy of the latitude values before passing to the
subroutine.

The fix is simple: we now just ALLOCATE the Arr1D array and copy
values from HcoState%GRID instead of just pointing to it.

Also: removed a line of debug output in hcoi_gc_main_mod.F90.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>