Initial modifications for making realclean except RRTMG
We have started making the initial modifications to allow
users to "realclean" everything except the RRTMG code.  Because
the RRTMG code takes a long time to compile, this will allow
for faster debugging and testing.

The basic idea is that each subdirectory Makefile will now
remove all of the objects, modules, and libraries that were
created for that subdirectory.  The work is ongoing, and a
few kinks still need to be worked out.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>