Modifications for MERRA2 met fields #3
Continue adding modifications for MERRA2 met fields to the rest of
the routines in GeosCore, in HEMCO/Core, and HEMCO/Extensions.

Fixed an error in the Makefile_header.mk: The headerinfo target
was left activated; we have now commented that out.

Also note: we have hardwired the location of the MERRA2 data files
for testing.  This hardwiring will be removed in the next commit,
as we will start putting MERRA2 data in the proper met field path.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>