Manually specify hourly interval for diagnostics output for testing
The GEOSCHEM_diagnostics collection is defined with the call to
DiagnCollection_Create in diagnostics_mod. Prior to this commit, the
output frequency (passed as integers deltaYMD and deltaHMS) was set to
'END' since that is the value specified in HEMCO_Config.rc for DiagnFreq.
This setting is supposed to result in diagnostics output only at the end
of the run.  However, using 'End' does not currently work properly for
GEOS-Chem diagnostics and no file was being generated for a 1-hr
simulation.

To temporarily correct this issue, deltaYMD and deltaHMS are now passed
to DiagnCollection_Create as 00000000 and 010000, which result in hourly
diagnostics output.

Also in this commit:
	- comment out ewl debugging code in HEMCO/Core/hcoio_diagn_mod

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>