Fix issues detected by Gfortran with debugging options on
HEMCO/Extensions/hcox_seasalt_mod.F90: We have removed an extraneous
assignment statement for SS_DEN.  This was placed before SS_DEN had
been allocated, thus causing a segfault.

HEMCO/Extensions/hcox_gfed_mod.F90: Moved the WHERE( HUMTROP > 1 ) block
into the IF ( IsGFED3 ) section.  HUMTROP is only defined if GFED3 is
used, so it should not be referenced outside of the IF ( IsGFED3 )
section.

HEMCO/Core/hco_interp_mod.F90: Gfortran does not like the WHERE statement
with different array sizes.  For now, we have rewritten this as a DO
loop that is blocked off with #ifdef LINUX_FORTRAN.  In a future commit,
we will verify that the DO-loop replacement provides identical results
with IFORT.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>