Force HEMCO to consider masks on all CPUs
HEMCO defines a rectilinear, lat-lon box to determine if a mask
fully covers the domain of a specific CPU. As a result, masks are
always considered when running HEMCO in an ESMF environment.
However, the method used caused problems during diagnostics, as
it would incorrectly show some masks as fully covering a domain
even when the emissions calculation would not use this fact. To
aid in interpretation, mask coverage precalculation is now
completely disabled when HEMCO is running in an ESMF environment.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>